Update reliabilityFunctions.R
Yo. I tried your function out on learning to look at faces and Gabi figured out that the function specifies the working directory. I'm pushing this change to suggest removing that so that anyone can run it for whatever working directory they want. You don't have to change it, but I figured it would be good to try it out on a different study :)